Chat: split tool markdown formatter into focused partials
